refactor(Conversation): create useFilesUploadDropzone hook
