+++ server/config/client.config.ts
@@ -52,6 +52,7 @@ export function generateConfig(params: ConfigGeneratorParams, env: Env) {
       APPLOCK_SCHEDULED_TIMEOUT: env.FEATURE_APPLOCK_SCHEDULED_TIMEOUT
         ? Number(env.FEATURE_APPLOCK_SCHEDULED_TIMEOUT)
         : null,
+      ENABLE_CELLS: env.FEATURE_ENABLE_CELLS == 'true',
       CHECK_CONSENT: env.FEATURE_CHECK_CONSENT != 'false',
       CONFERENCE_AUTO_MUTE: env.FEATURE_CONFERENCE_AUTO_MUTE == 'true',
       ENABLE_IN_CALL_REACTIONS: env.FEATURE_ENABLE_IN_CALL_REACTIONS == 'true',
