+++ src/script/components/Conversation/Conversation.tsx
@@ -45,6 +45,7 @@ import {formatBytes} from 'Util/util';
 import {ConversationFileDropzone} from './ConversationFileDropzone/ConversationFileDropzone';
 import {useReadReceiptSender} from './hooks/useReadReceipt';
 import {ReadOnlyConversationMessage} from './ReadOnlyConversationMessage';
+import {useFilesUploadDropzone} from './useFilesUploadDropzone/useFilesUploadDropzone';
 import {checkFileSharingPermission} from './utils/checkFileSharingPermission';
 
 import {ConversationState} from '../../conversation/ConversationState';
@@ -461,13 +462,22 @@ export const Conversation = ({
     [addReadReceiptToBatch, repositories.conversation, repositories.integration, updateConversationLastRead],
   );
 
+  const isCellsEnabled = Config.getConfig().FEATURE.ENABLE_CELLS;
+
+  const {getRootProps, getInputProps, open, isDragAccept} = useFilesUploadDropzone({
+    isTeam: inTeam,
+    cellsRepository: repositories.cells,
+  });
+
   return (
     <ConversationFileDropzone
-      inTeam={inTeam}
-      isCellsEnabled={false}
+      isDragAccept={isDragAccept}
+      isCellsEnabled={isCellsEnabled}
       isConversationLoaded={isConversationLoaded}
       activeConversationId={activeConversation?.id}
       onFileDropped={checkFileSharingPermission(uploadDroppedFiles)}
+      rootProps={getRootProps()}
+      inputProps={getInputProps()}
     >
       {activeConversation && (
         <>
@@ -547,7 +557,7 @@ export const Conversation = ({
                 onShiftTab={() => setMsgElementsFocusable(false)}
                 uploadDroppedFiles={uploadDroppedFiles}
                 uploadImages={uploadImages}
-                uploadFiles={uploadFiles}
+                uploadFiles={isCellsEnabled ? () => open() : uploadFiles}
               />
             ))}
