+++ src/script/components/Conversation/useFilesUploadDropzone/fileValidation/fileValidation.ts
