+++ src/script/components/InputBar/InputBar.test.tsx
@@ -19,7 +19,7 @@
 
 import {act, fireEvent, render, waitFor} from '@testing-library/react';
 
-import {FileWithPreview} from 'Components/Conversation/useFiles/useFiles';
+import {FileWithPreview} from 'Components/Conversation/useFilesUploadState/useFilesUploadState';
 import {InputBar} from 'Components/InputBar/index';
 import {withTheme} from 'src/script/auth/util/test/TestUtil';
 import {Config} from 'src/script/Config';
